validatePhoneNumber => isValidPhoneNumber
